Can bundle exe and dll files needed by exe
However, exe also need resource files such as png files to work properly. I am going to finish that soon.
