+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "install": "node init.js",
     "test": "node scripts/out/test/all.test.js",
-    "bundle": "node scripts/out/main/bundle.js"
+    "bundle": "node scripts/out/main/bundle.js --release-path Calculator/Release/ --resource-path Calculator/res/main/"
   },
   "repository": {
     "type": "git",
