+++ scripts/src/main/bundle.ts
@@ -1,9 +1,20 @@
 "use strict";
 
 import fs from "fs";
+import { makeSureCleanDirExists } from "./fs_ext.js";
 
-function makeSureDirExistsSync(path: string): void {
-    if (!fs.existsSync(path)) fs.mkdirSync(path);
-}
-
-makeSureDirExistsSync("out/");
+if (process.argv.length !== 6 || process.argv[2] !== "--release-path" || process.argv[4] !== "--resource-path")
+    throw new Error(
+        "Invalid Command Line!   Usage: node scripts/out/main/bundle.js --release-path <path> --resource-path <path>"
+    );
+const outPath = "out/";
+makeSureCleanDirExists(outPath);
+fs.readdir(process.argv[3], (err, files) => {
+    if (err) throw err;
+    for (const file of files) {
+        if (!(file.endsWith(".exe") || file.endsWith(".dll"))) continue;
+        fs.copyFile(`${process.argv[3]}${file}`, `${outPath}${file}`, err => {
+            if (err) throw err;
+        });
+    }
+});
