--- scripts/src/main/synthetic_fs.ts
@@ -1,5 +1,0 @@
-"use strict";
-
-export function getFileName(path: string): string {
-    return path.substring(path.lastIndexOf("/") + 1);
-}
